+++ src/main/java/com/yushan/backend/dto/NovelSearchRequestDTO.java
@@ -1,11 +1,22 @@
 package com.yushan.backend.dto;
 
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 @Data
-@EqualsAndHashCode(callSuper = true)
-public class NovelSearchRequestDTO extends NovelPaginationRequestDTO {
+public class NovelSearchRequestDTO {
+    
+    @Min(value = 0, message = "Page number must be >= 0")
+    private Integer page = 0;
+    
+    @Min(value = 1, message = "Page size must be >= 1")
+    @Max(value = 100, message = "Page size must be <= 100")
+    private Integer size = 10;
+    
+    private String sort = "createTime";
+    
+    private String order = "desc";
     
     private Integer categoryId;
     private String status;
@@ -13,12 +24,18 @@ public class NovelSearchRequestDTO extends NovelPaginationRequestDTO {
     private String authorName;
     
     public NovelSearchRequestDTO() {
-        super();
+        this.page = 0;
+        this.size = 10;
+        this.sort = "createTime";
+        this.order = "desc";
     }
     
     public NovelSearchRequestDTO(Integer page, Integer size, String sort, String order, 
                                Integer categoryId, String status, String search, String authorName) {
-        super(page, size, sort, order);
+        this.page = page != null ? page : 0;
+        this.size = size != null ? size : 10;
+        this.sort = sort != null ? sort : "createTime";
+        this.order = order != null ? order : "desc";
         this.categoryId = categoryId;
         this.status = status;
         this.search = search;
@@ -39,5 +56,13 @@ public class NovelSearchRequestDTO extends NovelPaginationRequestDTO {
     
     public boolean hasAuthorFilter() {
         return authorName != null && !authorName.trim().isEmpty();
+    }
+    
+    public boolean isAscending() {
+        return "asc".equalsIgnoreCase(order);
+    }
+    
+    public boolean isDescending() {
+        return "desc".equalsIgnoreCase(order);
     }
 }
